requirements.txt
tensorflow-gpu == 1.12.0
matplotlib >= 3.0.3
numpy == 1.14.5
pandas >= 0.24.1
plotly >= 3.7.0
SimpleITK >= 1.2.0
scikit-image >= 0.4.12
scipy >= 1.1.0
pymia == 0.2.1
torch